Set minimum and maximum values for all parameters.

Generates a dataframe defining the minimum and maximum values of each parameter. This dataframe of minimums and maximums is then turned into a list if scalers (single numbers) for parameters that do not vary and and vectors for numbers that do vary by the function param_seqs(). A dataframe of all possible parameter combinatnsion (aka, a parameter grid) is then made with param_grid().

param_ranges(
  scenario = NA,
  figure = NA,
  gamma. = c(5, 5),
  c. = c(1, 1),
  K.bc. = c(1000, 1000),
  K.bk. = c(10000, 10000),
  K.wg. = c(900, 900),
  S.w.mg. = c(0.8, 0.8),
  S.w.mp. = c(0.8, 0.8),
  S.w.fg. = c(0.8, 0.8),
  S.w.fp. = c(0.8, 0.8),
  S.m.mg. = c(0.9, 0.9),
  S.m.mp. = c(0.8, 0.8),
  S.m.fg. = c(0.9, 0.9),
  S.m.fp. = c(0.8, 0.8),
  R.base.rate. = c(1.8, 1.8),
  R.hab.effect. = c(0.5, 0.5),
  co. = c(1, 1),
  f. = c(0.5, 0.5),
  S.b.mc. = c(0.95, 0.95),
  S.b.mk. = c(0.85, 0.85),
  S.b.md. = c(0.8, 0.8),
  S.b.fc. = c(0.95, 0.95),
  S.b.fk. = c(0.85, 0.85),
  S.f.mc. = c(0.8, 0.8),
  S.f.mk. = c(0.75, 0.75),
  S.f.md. = c(0.8, 0.8),
  S.f.fc. = c(0.8, 0.8),
  S.f.fk. = c(0.75, 0.75),
  S.y.mc. = c(0.8, 0.8),
  S.y.mk. = c(0.75, 0.75),
  S.y.fc. = c(0.8, 0.8),
  S.y.fk. = c(0.75, 0.75),
  verbose = TRUE
)

Arguments

scenario
Different habitat scenarios defined by Runge & Marra 2004. In particular sets carrying capacity for breeding source habitat (K.bc) and winter good habitat (K.wg) If set overides other varibles that may have been set using other arguements.

figure
Generate figure. Currently only figures 28.3 and 28.4 are implemented

gamma.
Strength of male dominance.

c.
Strength of winter to breeding season carry over.

K.bc.
Carrying capacity (K) of source (high quality) breeding habitat.

K.bk.
Carrying capacity (K) of sink (low quality) breeding habitat.

K.wg.
Carrying capacity (K) of winter good (high quality) habitat.

S.w.mg.
Survival (S) during winter (.w.) of males (.m) in good habitat (g). Note that there is no age structure in winter.

S.w.mp.
Survival (S) during winter (.w.) of males (.m) in poor habitat (p).

S.w.fg.
Survival (S) during winter (.w.) of females (.f) in good habitat (g).

S.w.fp.
Survival (S) during winter (.w.) of females (.f) in good habitat (p).

S.m.mg.
Survival (S) during spring migration (.m.) of males (.m) originating from good (g) winter habitat.

S.m.mp.
Survival (S) during spring migration (.m.) of males (.m) originating from poor (p) winter habitat.

S.m.fg.
Survival (S) during spring migration (.m.) of females (.f) originating from good (g) winter habitat.

S.m.fp.
Survival (S) during spring migration (.m.) of females (.m) originating from poor (p) winter habitat.

R.base.rate.
Fecundity base rate

R.hab.effect.
Effect on fecundity of a pair breeding in poor habitat

co.
Magnitude of carry over effect; c = 1 is no carry over

f.
Sex ratio

S.b.mc.
Survival (S) during breeding (.b.) of males (.m) in source (c) habitat (source = good habitat)

S.b.mk.
Survival (S) during breeding (.b.) of males (.m) in sink (k) habitat (sink = poor habitat)

S.b.md.
Survival (S) during breeding (.b.) of males (.m) with drain (d) status (drain = floater or unpaired)

S.b.fc.
Survival (S) during breeding (.b.) of females (.f) in source (c) habitat (source = good habitat)

S.b.fk.
Survival (S) during breeding (.b.) of females (.f) in sink (k) habitat (sink = poor habitat)

S.f.mc.
Survival (S) during fall (.f.) migration of males (.m) originating from source (c) habitat

S.f.mk.
Survival (S) during fall (.f.) migration of males (.m) originating from sink (k) habitat

S.f.md.
Survival (S) during fall (.f.) migration of males (.m) with drain (d) status (aka floater, unpaired)

S.f.fc.
Survival (S) during fall (.f.) migration of females (.f) originating from source (c) habitat

S.f.fk.
Survival (S) during fall (.f.) migration of females (.f) originating from sink (k) habitat

S.y.mc.
Survival (S) during fall migration of young (.y.) males (.m) originating from source (c) habitat

S.y.mk.
Survival (S) during fall migration of young (.y.) males (.m) originating from sink (k) habitat

S.y.fc.
Survival (S) during fall migration of young (.y.) females (.f) originating from source (c) habitat

S.y.fk.
Survival (S) during fall migration of young (.y.) females (.f) originating from sink (k) habitat

verbose
T/F return messages

Value

param.ranges 30 x 2 dataframe of the min. and max. parameter values that will be considered. Often values are fixed and do not vary and so are listed as both min and max.

Details

Default values are based on Runage and Marra (2004). The arguments "scenario" and "figure" can be used to set up the parameters for specific figures from Runge and Marra (2004).

References

Runge, MC and PP Marra. 2004. Modeling seasonal interactions in the population dynamics of migratory birds. In Greenberg, R and PP Marra, eds. Birds of two worlds: The ecology & evolution of migration. Johns Hopkins University Press, Baltimore.

Examples

# Set parameters for the winter limited scenario in original paper (Runge & Marra 2004)
param.ranges <- param_ranges(scenario = "winter.limited")

#> Setting breeding 'source' carrying capacity K.bc to 800
#> Setting winter   'good'   carrying capacity K.wg to 484

head(param.ranges)

#>            min     max
#> gamma      5.0     5.0
#> co.        1.0     1.0
#> K.bc     800.0   800.0
#> K.bk   10000.0 10000.0
#> K.wg     485.0   485.0
#> S.w.mg     0.8     0.8

# Parameters to remake Figure 28.3 in original paper (Runge & Marra 2004)
param.ranges <- param_ranges(figure = 28.3)
param.ranges[c("K.bc","K.wg"), ]

#>      min  max
#> K.bc   1 1000
#> K.wg   1 1000

# Setting carrying capacities to 0 causes errors when running model
## param_ranges() therefore automatically changes to prevent this bug
param.ranges <- param_ranges(K.bc. = c(0,100))

#> One of your carrying capacities is set to 0, which due to an unresolved bug can cause problems.  This value has been set to 1.  Sorry.

param.ranges[c("K.bc","K.wg"), ]

#>      min max
#> K.bc   1 100
#> K.wg 900 900

# Set up different scenarios
## this is similar to data for figure 28.5, but without gamma varying
winter.lim <- param_ranges(scenario = "winter")

#> Setting breeding 'source' carrying capacity K.bc to 800
#> Setting winter   'good'   carrying capacity K.wg to 484

intermed <- param_ranges(scenario = "intermediate")

#> Setting breeding 'source' carrying capacity K.bc to 224
#> Setting winter   'good'   carrying capacity K.wg to 580

summer.lim <- param_ranges(scenario = "summer")

#> Setting breeding 'source' carrying capacity K.bc to 205
#> Setting winter   'good'   carrying capacity K.wg to 900

# Show data
## Note only 1st 6 rows shown; delete "[1:6,]" to show all.
data.frame(winter.lim, intermed, summer.lim)[1:6,]

#>            min     max   min.1   max.1   min.2   max.2
#> gamma      5.0     5.0     5.0     5.0     5.0     5.0
#> co.        1.0     1.0     1.0     1.0     1.0     1.0
#> K.bc     800.0   800.0   224.0   224.0   205.0   205.0
#> K.bk   10000.0 10000.0 10000.0 10000.0 10000.0 10000.0
#> K.wg     485.0   485.0   580.0   580.0   900.0   900.0
#> S.w.mg     0.8     0.8     0.8     0.8     0.8     0.8